End point

+++ softdesk/issues/views.py
@@ -14,11 +14,12 @@ class IssuesViewset(viewsets.ModelViewSet):
     # form_class = IssuesForm
 
     def get_queryset(self):
-        return Issues.objects.all()
+        pk = self.kwargs['pk']
+        return Issues.objects.filter(project_id=pk)
         
-    def create(self, request, project_pk=None):
+    def create(self, request, pk=None):
         data = request.data.copy()
-        data['project_id']=Project.objects.get(pk=project_pk).project_id
+        data['project_id']=Project.objects.get(pk=pk).project_id
         data['author_user_id']=request.user.user_id
         data['assignee_user_id']=User.objects.get(
             first_name=request.POST.get('assignee')
@@ -36,23 +37,43 @@ class IssuesViewset(viewsets.ModelViewSet):
     def update(self, request, project_pk=None, pk=None):
         try:
             issue = Issues.objects.get(pk=pk)
-        except issue.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+            
+            if not issue.project_id.project_id == project_pk:
+                return Response(
+                    "le problème n'appartient pas à ce projet",
+                    status=status.HTTP_401_UNAUTHORIZED
+                )
 
-        serializer = IssuesSerializer(issue, data=request.data, partial=True)
-        # setting raise_exception=True in the serializer's is_valid method will raise exception on error. So you don't have implement extra logics
-        serializer.is_valid(raise_exception=True)
-        serializer.save(user=request.user)
-        return Response(serializer.data)
-    
+            serializer = IssuesSerializer(issue, data=request.data, partial=True)
+            serializer.is_valid(raise_exception=True)
+            serializer.save(user=request.user)
+            return Response(serializer.data)
+        
+        except issue.DoesNotExist:
+            return Response(
+                "le problème n'existe pas",
+                status=status.HTTP_404_NOT_FOUND
+                )
+
     def delete(self, request, project_pk=None, pk=None):
         try:
             issue = Issues.objects.get(pk=pk)
+
+            if not issue.project_id.project_id == project_pk:
+                return Response(
+                    "le problème n'appartient pas à ce projet",
+                    status=status.HTTP_401_UNAUTHORIZED
+                )
+            
             issue.delete()
-            return Response(status=status.HTTP_200_OK)
+            return Response(
+                "Le problème à bien été supprimé",
+                status=status.HTTP_200_OK,
+                )
         except issue.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+            return Response(
+                "Le problème n'existe pas",
+                status=status.HTTP_404_NOT_FOUND
+                )
         except:
             return Response(issue.errors, status=status.HTTP_204_NO_CONTENT)
- 
-
